[PSCORE] Merge embeddings from local(hdfs) to a binary file
