+++ python/paddle/distributed/fleet/base/fleet_base.py
@@ -824,160 +824,6 @@ class Fleet(object):
         self._runtime_handle._save_persistables(executor, dirname, main_program,
                                                 mode)
 
-    def merge_embedding_shards_to_binary(self, input_path, output_path,
-                                         **kwargs):
-        """
-        get embedding shards from local files and merge them into a single file with binary format
-
-        Args:
-            input_path(str): local input path
-            output_path(str): local output path
-            kwargs(dict): user defined properties
-                          feasign_nums(int): initialize number of feasigns for each shard file (initialize reserve vector space)
-                          hadoop_fs_name(str): hadoop fs name
-                          hadoop_fs_ugi(str): hadoop fs ugi
-                          hadoop_home(str): hadoop home, default is "$HADOOP_HOME"
-
-        Examples:
-            .. code-block:: text
-              import paddle
-              paddle.enable_static()
-              import paddle.distributed.fleet as fleet
-
-              fleet.init()
-              exe = paddle.static.Executor(paddle.CPUPlace())
-              fleet.merge_embedding_shards_to_binary("input_path", "output_path")
-        """
-        feasign_nums = kwargs.get("feasign_nums", 1000)
-        hadoop_fs_name = kwargs.get("hadoop_fs_name", None)
-        hadoop_fs_ugi = kwargs.get("hadoop_fs_ugi", None)
-        hadoop_home = kwargs.get("hadoop_home", "$HADOOP_HOME")
-        upload_flag = False
-
-        print("going to merge embedding shards to binary!")
-
-        if os.path.isdir("output_embedding/"):
-            shutil.rmtree("output_embedding/", True)
-
-        # get files from hdfs
-        if input_path.startswith("hdfs:") or input_path.startswith("afs:"):
-            print("going to download embedding shards from hdfs!")
-            configs = {
-                "fs.default.name": hadoop_fs_name,
-                "hadoop.job.ugi": hadoop_fs_ugi
-            }
-            client = HDFSClient(hadoop_home, configs)
-            client.download(input_path, "output_embedding/")
-            input_path = "output_embedding/"
-
-        if not os.path.exists(input_path):
-            raise ValueError \
-                ("{} is not exist, pleast confirm your input_path.".format(input_path))
-
-        if output_path.startswith("hdfs:") or output_path.startswith("afs:"):
-            tmp_output_path = output_path
-            output_path = "merged_embedding/"
-            upload_flag = True
-
-        if os.path.isdir(output_path):
-            shutil.rmtree(output_path, True)
-        os.mkdir(output_path)
-
-        # get var names from input path
-        var_names = []
-        for f in os.listdir(input_path):
-            f_path = os.path.join(input_path, f)
-        if os.path.isdir(f_path) and f.endswith(".shard"):
-            var_names.append(f[:-6])
-
-        # merge embedding shards for each var
-        for varname in var_names:
-            shard_txt = os.path.join(input_path, "{}.shard".format(varname))
-            save_path = os.path.join(output_path, varname)
-
-            if not os.path.exists(shard_txt):
-                raise ValueError \
-                    ("{} is not exist, pleast confirm your input_path.".format(shard_txt))
-
-            if os.path.exists(save_path):
-                raise ValueError("{} is exist, pleast delete.".format(
-                    save_path))
-
-            print("searching Param/Meta from {} and will merge to {}".format(
-                shard_txt, save_path))
-
-            shards, param_dim = self.get_distributed_shard(shard_txt, varname)
-            sharding_merge = paddle.fluid.core.ShardingMerge()
-            sharding_merge.merge(shards, [feasign_nums for i in shards],
-                                 save_path, param_dim)
-
-            print("save {} with {} shards to {}".format(varname,
-                                                        len(shards), save_path))
-
-        # upload files to hdfs
-        if upload_flag:
-            print("going to upload merged embedding shards to hdfs!")
-            configs = {
-                "fs.default.name": hadoop_fs_name,
-                "hadoop.job.ugi": hadoop_fs_ugi
-            }
-            client = HDFSClient(hadoop_home, configs)
-
-            if not client.is_exist(tmp_output_path):
-                client.mkdirs(tmp_output_path)
-            client.upload(
-                output_path, tmp_output_path, multi_processes=5, overwrite=True)
-
-        print("merge embedding shards done!")
-
-    def get_distributed_shard(self, shard_dirname, shard_varname):
-        def get_meta(shard_meta):
-            varname = None
-            param_dim = -1
-            row_names = None
-            row_dims = None
-
-            with open(shard_meta, "r") as rb:
-                for line in rb:
-                    line = line.strip()
-                    if line.startswith("param="):
-                        varname = line.split("=")[1]
-                    if line.startswith("row_name"):
-                        row_names = line.split("=")[1]
-                    if line.startswith("row_dims"):
-                        row_dims = line.split("=")[1]
-
-                param_dim = row_dims.split(",")[row_names.split(",").index(
-                    "Param")]
-                param_dim = int(param_dim)
-
-            if varname is None or param_dim == -1:
-                raise ValueError \
-                    ("can not get right information from {}".format(shard_meta))
-
-            return (varname, param_dim)
-
-        def get_shard():
-            shards = []
-            for f in os.listdir(shard_dirname):
-                if f.startswith(shard_varname) and f.endswith(".txt"):
-                    shards.append(os.path.join(shard_dirname, f))
-            return shards
-
-        shards = get_shard()
-        if len(shards) == 0:
-            raise ValueError(
-                "can't find embedding shard txt files, please check")
-
-        meta_txt = os.path.join(shard_dirname,
-                                "{}.block0.meta".format(shard_varname))
-        meta_varname, param_dim = get_meta(meta_txt)
-
-        if meta_varname != shard_varname:
-            raise ValueError("get_distributed_shard meta error, please check.")
-
-        return shards, param_dim
-
     def shrink(self, threshold):
         self._runtime_handle._shrink(threshold)
